requirements.txt
aiohttp==3.13.3
anyio==4.12.1
attrs==25.4.0
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
dataclasses-json==0.6.7
distro==1.9.0
fastapi==0.128.0
frozenlist==1.8.0
groq==0.37.1
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
idna==3.11
jsonpatch==1.33
jsonpointer==3.0.0
langchain==1.2.6
langchain-groq==1.1.1
langchain-classic==1.0.1
langchain-core==1.2.7
langchain-text-splitters==1.1.0
langchain-community==0.4.1
marshmallow==3.26.2
multidict==6.7.0
numpy==2.2.6
opencv-python==4.12.0.88
orjson==3.11.5
packaging==25.0
pdf2image==1.17.0
Pillow==12.1.0
pillow-avif-plugin==1.5.2
pytesseract==0.3.13
python-dotenv==1.2.1
python-multipart==0.0.21
PyYAML==6.0.3
requests==2.32.5
setuptools==65.5.0
SQLAlchemy==2.0.45
starlette==0.50.0
tenacity==9.1.2
typing_extensions==4.15.0
uvicorn==0.40.0
uuid_utils==0.13.0
xxhash==3.6.0
yarl==1.22.0
zstandard==0.25.0
